tests/wpt/mozilla/tests/bluetooth/getCharacteristics/correct-characteristics.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => service.getCharacteristics(body_sensor_location.name))
    .then(characteristics => Promise.all([
        characteristics[0].readValue(),
        characteristics[1].readValue()]))
    .then(value_arrays => assert_array_equals(value_arrays, ['1', '2']));
}, 'Find characteristics with UUID in service.');
</script>
